+++ src/pygaps/calculations/alphas.py
@@ -125,7 +125,7 @@ def alpha_s(isotherm, reference_isotherm, reference_area=None,
     # Get adsorbate properties
     adsorbate = Adsorbate.find(isotherm.adsorbate)
     molar_mass = adsorbate.molar_mass()
-    liquid_density = adsorbate.liquid_density(isotherm.t_iso)
+    liquid_density = adsorbate.liquid_density(isotherm.temperature)
 
     # Read data in
     loading = isotherm.loading(branch='ads',