+++ src/pygaps/calculations/dr_da_plots.py
@@ -175,8 +175,8 @@ def da_plot(isotherm, exp=None, limits=None, verbose=False):
     # Get adsorbate properties
     adsorbate = Adsorbate.find(isotherm.adsorbate)
     molar_mass = adsorbate.molar_mass()
-    liquid_density = adsorbate.liquid_density(isotherm.t_iso)
-    iso_temp = isotherm.t_iso
+    liquid_density = adsorbate.liquid_density(isotherm.temperature)
+    iso_temp = isotherm.temperature
 
     # Read data in
     loading = isotherm.loading(branch='ads',